conda-requirements.txt
pandas
bokeh
dill
flask
jinja2
pip
simplejson
alpha_vantage.timeseries
alpha_vantage.techindicators
matplotlib.pyplot